Allow a function to have a retry-spec
PSF.with_retry_spec(): Add wrapper which in Python just runs the given
function with the given args, but indicates the intent that it have
the given retry-spec attached to it when converted to a Step Function
state machine.

summarise(): Give the augment_summary() call a retry-spec.

No change to (Python) behaviour.